+++ tests/contrib/test_dropbox.py
@@ -83,17 +83,16 @@ def test_context_local(make_app):
         assert request.headers["Authorization"] == "Bearer app2"
 
 
-def _test_app_redirect(app):
+def app_redirect_location(app):
     with app.test_client() as client:
         resp = client.get("/dropbox", base_url="https://a.b.c", follow_redirects=False)
     assert resp.status_code == 302
-    location = URLObject(resp.headers["Location"])
-    return location
+    return URLObject(resp.headers["Location"])
 
 
 def test_default_redirect_params(make_app):
     app = make_app("foo", "bar")
-    query_dict = _test_app_redirect(app).query_dict
+    query_dict = app_redirect_location(app).query_dict
     assert isinstance(query_dict.pop("state"), string_types)
     assert query_dict == {
         "client_id": "foo",
@@ -104,19 +103,19 @@ def test_default_redirect_params(make_app):
 
 def test_force_reapprove(make_app):
     app = make_app("foo", "bar", force_reapprove=True)
-    assert _test_app_redirect(app).query_dict["force_reapprove"] == "true"
+    assert app_redirect_location(app).query_dict["force_reapprove"] == "true"
 
 
 def test_disable_signup(make_app):
     app = make_app("foo", "bar", disable_signup=True)
-    assert _test_app_redirect(app).query_dict["disable_signup"] == "true"
+    assert app_redirect_location(app).query_dict["disable_signup"] == "true"
 
 
 def test_require_role(make_app):
     app = make_app("foo", "bar", require_role="work")
-    assert _test_app_redirect(app).query_dict["require_role"] == "work"
+    assert app_redirect_location(app).query_dict["require_role"] == "work"
 
 
 def test_offline(make_app):
     app = make_app("foo", "bar", offline=True)
-    assert _test_app_redirect(app).query_dict["token_access_type"] == "offline"
+    assert app_redirect_location(app).query_dict["token_access_type"] == "offline"